pyproject.toml
[tool.poetry]
name = "mlflow-docker-serve"
version = "0.1.0"
description = "An MLFlow plugin that allows one-click MLFlow deployments using docker"
authors = ["Aucke Bos <aucke.bos97@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "mlflow_docker_serve"}]

[tool.poetry.dependencies]
python = "^3.9"
mlflow = "^2.6.0"
prefect = "^2.11.3"
pymongo = "^4.4.1"
python-dotenv = "^1.0.0"
kink = "^0.6.6"
pandas = "^2.0.3"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.7.0"
pre-commit = "^3.3.3"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
